refactor: withAuthPage로 인증/인가 구현
